# lib/api-zod/src/generated/types/contactMessageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessageInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  topic: string;
  order?: string;
  /** @minLength 1 */
  message: string;
}